Fb fix hgi (#40)
* Fixed gsub issue

* Wip, fixing the h-ratios, added some documentation and output generation

* Made some modifications to the contrained glm

* Use data[t] instead of data[0]

* Updated constraints / looping over observations

* Minor fixes to the glm function

* Pre restructure commit

* Added some fixes, included multi-node cluster script (not yet entirely working)

* Added some scripts to run the configurations on the cluster

* Added missing test, fixed failing test

* Added learning curve plotter and fixed CV score for DOSL

* Fixed some warnings, added delta to mu.eta and fixed some specs

* Added dependency

* Minor fixes, reverted the cross_df becuase of cluster dependencies

* Added some more specs

* Loop over different blocks from N

* Remove the dev resids func

* Fixes for h-ratios, general updates to code and visualization

* Move residuals function

* Added test cases

* Updated with C functions for constrained glm

* Updating the oos estimator, so it can be updated

* Updated some internals of the one step estimator. Next ->> make it online.

* Updated specs / OOS implementation

* Updated documentation

* Removed condensier line

* Added specs for interventioneffectcalculator

* Added better multicore seeds

* Added svm / randomforest

* Increase theshhold

+++ inst/bash/run_on_cluster.sh
@@ -1,14 +1,20 @@
 #!/usr/bin/env sh
-server_name=peregrine
-server_name=peregrine-interactive
-scp -r * $server_name:~/osl
 #rsync -ue ssh ./ peregrine:~/osl
-if [[ $server_name != 'peregrine-interactive' ]]; then
-  ssh $server_name -t 'cd osl; srun --ntasks=1 --cpus-per-task=22 --time=00:30:00 --partition=nodes Rscript run.R'
-else
+
+if [ $# -eq 0 ]; then
+  server_name=peregrine-interactive
+  scp -r * $server_name:~/osl
   ssh $server_name -t 'rm -rf ~/tmp; mkdir -p ~/tmp/osl/; cd osl; Rscript run.R'
-  #ssh $server_name -t 'rm -r ~/tmp; mkdir ~/tmp/osl/; cd osl; R'
+  rm -r /tmp/osl/
+  mkdir -p /tmp/osl/
+  scp -r $server_name:~/tmp/* /tmp/osl/
+else
+  server_name=peregrine
+  scp -r * $server_name:~/osl
+  # 4 is the number of configurations
+  for i in $(seq 1 4);
+  do
+    ssh $server_name -t "cd osl; sbatch -o ~/tmp/$i-output.out -e ~/tmp/$i-error.out --job-name=cfg-$i --ntasks=1 --cpus-per-task=23 --time=16:00:00 --partition=regular --mail-user=peregrine@compsy.nl --mail-type=all --wrap=\"Rscript run.R $i\""
+  done
 fi
-rm -r /tmp/osl/
-mkdir -p /tmp/osl/
-scp -r $server_name:~/tmp/* /tmp/osl/
+
